+++ package.yaml
@@ -44,7 +44,7 @@ library:
     - csrc/swepdate.c
     - csrc/sweph.c
     - csrc/swephlib.c
-    - csrc/swegrav.c
+    - csrc/gravgroup.c
   include-dirs: csrc
   install-includes:
     - csrc/swedate.h
@@ -58,7 +58,7 @@ library:
     - csrc/sweph.h
     - csrc/swephexp.h
     - csrc/swephlib.h
-    - csrc/swegrav.h
+    - csrc/gravgroup.h
 
 tests:
   swiss-ephemeris-test:
