+++ swiss-ephemeris.cabal
@@ -4,7 +4,7 @@ cabal-version: 1.12
 --
 -- see: https://github.com/sol/hpack
 --
--- hash: 28a46b1135a8e3c3bcb3366cebc859774b8e2efaad94aa07758cdcf681bc2bdf
+-- hash: e634b9f9611fdf3c8ebfe9471e34ef3907a1db90eabbde3468a87e7353c0b2e8
 
 name:           swiss-ephemeris
 version:        1.3.0.0
@@ -52,7 +52,7 @@ library
       csrc/sweph.h
       csrc/swephexp.h
       csrc/swephlib.h
-      csrc/swegrav.h
+      csrc/gravgroup.h
   c-sources:
       csrc/swecl.c
       csrc/swedate.c
@@ -65,7 +65,7 @@ library
       csrc/swepdate.c
       csrc/sweph.c
       csrc/swephlib.c
-      csrc/swegrav.c
+      csrc/gravgroup.c
   build-depends:
       base >=4.10 && <4.16
   default-language: Haskell2010
